refactor : apply code review - separate test config
