+++ src/main/java/nextstep/auth/token/oauth2/github/GithubClient.java
@@ -5,7 +5,6 @@ import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
-import org.springframework.stereotype.Component;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
@@ -21,30 +20,32 @@ public class GithubClient implements GithubClientInterface {
     @Value("${github.url.profile}")
     private String profileUrl;
 
+    @Override
     public String getAccessTokenFromGithub(String code) {
         GithubAccessTokenRequest githubAccessTokenRequest = new GithubAccessTokenRequest(
-            code,
-            clientId,
-            clientSecret
+                code,
+                clientId,
+                clientSecret
         );
 
         HttpHeaders headers = new HttpHeaders();
         headers.add("Accept", MediaType.APPLICATION_JSON_VALUE);
 
         HttpEntity<MultiValueMap<String, String>> httpEntity = new HttpEntity(
-            githubAccessTokenRequest, headers);
+                githubAccessTokenRequest, headers);
         RestTemplate restTemplate = new RestTemplate();
 
         String accessToken = restTemplate
-            .exchange(tokenUrl, HttpMethod.POST, httpEntity, GithubAccessTokenResponse.class)
-            .getBody()
-            .getAccessToken();
+                .exchange(tokenUrl, HttpMethod.POST, httpEntity, GithubAccessTokenResponse.class)
+                .getBody()
+                .getAccessToken();
         if (accessToken == null) {
             throw new RuntimeException();
         }
         return accessToken;
     }
 
+    @Override
     public GithubProfileResponse getGithubProfileFromGithub(String accessToken) {
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", "token " + accessToken);
@@ -54,8 +55,8 @@ public class GithubClient implements GithubClientInterface {
 
         try {
             return restTemplate
-                .exchange(profileUrl, HttpMethod.GET, httpEntity, GithubProfileResponse.class)
-                .getBody();
+                    .exchange(profileUrl, HttpMethod.GET, httpEntity, GithubProfileResponse.class)
+                    .getBody();
         } catch (HttpClientErrorException e) {
             throw new RuntimeException();
         }
